refactor, move out exception handling
