+++ AppCheckCore/Sources/AppAttestProvider/Errors/GACAppAttestRejectionError.m
@@ -18,7 +18,7 @@
 
 #import "AppCheckCore/Sources/AppAttestProvider/Errors/GACAppAttestRejectionError.h"
 
-#import "AppCheckCore/Sources/Public/AppCheckCore/GACAppCheckErrorUtil.h"
+#import "AppCheckCore/Sources/Public/AppCheckCore/_GACAppCheckErrorUtil.h"
 
 @implementation GACAppAttestRejectionError
