+++ AppCheckCore/Sources/Core/APIService/NSURLSession+GACPromises.m
@@ -22,11 +22,11 @@
 #import "FBLPromises.h"
 #endif
 
-#import "AppCheckCore/Sources/Public/AppCheckCore/GACURLSessionDataResponse.h"
+#import "AppCheckCore/Sources/Public/AppCheckCore/_GACURLSessionDataResponse.h"
 
 @implementation NSURLSession (GACPromises)
 
-- (FBLPromise<GACURLSessionDataResponse *> *)gac_dataTaskPromiseWithRequest:
+- (FBLPromise<_GACURLSessionDataResponse *> *)gac_dataTaskPromiseWithRequest:
     (NSURLRequest *)URLRequest {
   return [FBLPromise async:^(FBLPromiseFulfillBlock fulfill, FBLPromiseRejectBlock reject) {
     [[self dataTaskWithRequest:URLRequest
@@ -35,7 +35,7 @@
                if (error) {
                  reject(error);
                } else {
-                 fulfill([[GACURLSessionDataResponse alloc]
+                 fulfill([[_GACURLSessionDataResponse alloc]
                      initWithResponse:(NSHTTPURLResponse *)response
                              HTTPBody:data]);
                }
