+++ AppCheckCore/Sources/DebugProvider/API/GACAppCheckDebugProviderAPIService.m
@@ -26,7 +26,7 @@
 #import "AppCheckCore/Sources/Public/AppCheckCore/GACAppCheckAPIService.h"
 
 #import "AppCheckCore/Sources/Core/GACAppCheckLogger+Internal.h"
-#import "AppCheckCore/Sources/Public/AppCheckCore/GACAppCheckErrorUtil.h"
+#import "AppCheckCore/Sources/Public/AppCheckCore/_GACAppCheckErrorUtil.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -37,7 +37,7 @@ static NSString *const kLimitedUseField = @"limited_use";
 
 @interface GACAppCheckDebugProviderAPIService ()
 
-@property(nonatomic, readonly) id<GACAppCheckAPIServiceProtocol> APIService;
+@property(nonatomic, readonly) id<_GACAppCheckAPIServiceProtocol> APIService;
 
 @property(nonatomic, readonly) NSString *resourceName;
 
@@ -45,7 +45,7 @@ static NSString *const kLimitedUseField = @"limited_use";
 
 @implementation GACAppCheckDebugProviderAPIService
 
-- (instancetype)initWithAPIService:(id<GACAppCheckAPIServiceProtocol>)APIService
+- (instancetype)initWithAPIService:(id<_GACAppCheckAPIServiceProtocol>)APIService
                       resourceName:(NSString *)resourceName {
   self = [super init];
   if (self) {
@@ -64,13 +64,13 @@ static NSString *const kLimitedUseField = @"limited_use";
   NSURL *URL = [NSURL URLWithString:URLString];
 
   return [self HTTPBodyWithDebugToken:debugToken limitedUse:limitedUse]
-      .then(^FBLPromise<GACURLSessionDataResponse *> *(NSData *HTTPBody) {
+      .then(^FBLPromise<_GACURLSessionDataResponse *> *(NSData *HTTPBody) {
         return [self.APIService sendRequestWithURL:URL
                                         HTTPMethod:@"POST"
                                               body:HTTPBody
                                  additionalHeaders:@{kContentTypeKey : kJSONContentType}];
       })
-      .then(^id _Nullable(GACURLSessionDataResponse *_Nullable response) {
+      .then(^id _Nullable(_GACURLSessionDataResponse *_Nullable response) {
         return [self.APIService appCheckTokenWithAPIResponse:response];
       });
 }
@@ -82,7 +82,7 @@ static NSString *const kLimitedUseField = @"limited_use";
   if (debugToken.length <= 0) {
     FBLPromise *rejectedPromise = [FBLPromise pendingPromise];
     [rejectedPromise
-        reject:[GACAppCheckErrorUtil errorWithFailureReason:@"Debug token must not be empty."]];
+        reject:[_GACAppCheckErrorUtil errorWithFailureReason:@"Debug token must not be empty."]];
     return rejectedPromise;
   }
 
@@ -100,7 +100,7 @@ static NSString *const kLimitedUseField = @"limited_use";
                             if (payloadJSON != nil) {
                               return payloadJSON;
                             } else {
-                              return [GACAppCheckErrorUtil JSONSerializationError:encodingError];
+                              return [_GACAppCheckErrorUtil JSONSerializationError:encodingError];
                             }
                           }];
 }
