+++ AppCheckCore/Sources/Public/AppCheckCore/GACAppCheckAPIService.h
@@ -14,52 +14,4 @@
  * limitations under the License.
  */
 
-#import <Foundation/Foundation.h>
-
-#import "GACAppCheckProvider.h"
-
-@class FBLPromise<Result>;
-@class GACURLSessionDataResponse;
-@class GACAppCheckToken;
-
-NS_ASSUME_NONNULL_BEGIN
-NS_SWIFT_NAME(AppCheckCoreAPIServiceProtocol)
-@protocol GACAppCheckAPIServiceProtocol <NSObject>
-
-@property(nonatomic, readonly) NSString *baseURL;
-
-- (FBLPromise<GACURLSessionDataResponse *> *)
-    sendRequestWithURL:(NSURL *)requestURL
-            HTTPMethod:(NSString *)HTTPMethod
-                  body:(nullable NSData *)body
-     additionalHeaders:(nullable NSDictionary<NSString *, NSString *> *)additionalHeaders;
-
-- (FBLPromise<GACAppCheckToken *> *)appCheckTokenWithAPIResponse:
-    (GACURLSessionDataResponse *)response;
-
-@end
-NS_SWIFT_NAME(AppCheckCoreAPIService)
-@interface GACAppCheckAPIService : NSObject <GACAppCheckAPIServiceProtocol>
-
-/**
- * The default initializer.
- * @param session The URL session used to make network requests.
- * @param baseURL The base URL for the App Check service, e.g.,
- * `https://firebaseappcheck.googleapis.com/v1`.
- * @param APIKey The Google Cloud Platform API key, if needed, or nil.
- * @param requestHooks Hooks that will be invoked on requests through this service.
- */
-- (instancetype)initWithURLSession:(NSURLSession *)session
-                           baseURL:(nullable NSString *)baseURL
-                            APIKey:(nullable NSString *)APIKey
-                      requestHooks:(nullable NSArray<GACAppCheckAPIRequestHook> *)requestHooks
-    NS_DESIGNATED_INITIALIZER;
-
-- (instancetype)init NS_UNAVAILABLE;
-
-- (FBLPromise<GACAppCheckToken *> *)appCheckTokenWithAPIResponse:
-    (GACURLSessionDataResponse *)response;
-
-@end
-
-NS_ASSUME_NONNULL_END
+#import "_GACAppCheckAPIService.h"
