+++ AppCheckCore/Sources/Public/AppCheckCore/GACAppCheckErrorUtil.h
@@ -14,59 +14,4 @@
  * limitations under the License.
  */
 
-#import <Foundation/Foundation.h>
-
-@class GACAppCheckHTTPError;
-
-NS_ASSUME_NONNULL_BEGIN
-
-extern NSString *const kGACAppCheckMissingRecaptchaSDKMessage NS_SWIFT_NAME(missingRecaptchaSDKMessage);
-
-void GACAppCheckSetErrorToPointer(NSError *error, NSError **pointer);
-
-@interface GACAppCheckErrorUtil : NSObject
-
-+ (NSError *)publicDomainErrorWithError:(NSError *)error;
-
-// MARK: - Internal errors
-
-+ (NSError *)cachedTokenNotFound;
-
-+ (NSError *)cachedTokenExpired;
-
-+ (NSError *)keychainErrorWithError:(NSError *)error;
-
-+ (GACAppCheckHTTPError *)APIErrorWithHTTPResponse:(NSHTTPURLResponse *)HTTPResponse
-                                              data:(nullable NSData *)data;
-
-+ (NSError *)APIErrorWithNetworkError:(NSError *)networkError;
-
-+ (NSError *)appCheckTokenResponseErrorWithMissingField:(NSString *)fieldName;
-
-+ (NSError *)appAttestAttestationResponseErrorWithMissingField:(NSString *)fieldName;
-
-+ (NSError *)JSONSerializationError:(NSError *)error;
-
-+ (NSError *)errorWithFailureReason:(NSString *)failureReason;
-
-+ (NSError *)unsupportedAttestationProvider:(NSString *)providerName;
-
-+ (NSError *)missingRecaptchaSDKError;
-
-// MARK: - App Attest Errors
-
-+ (NSError *)appAttestKeyIDNotFound;
-
-+ (NSError *)appAttestGenerateKeyFailedWithError:(NSError *)error;
-
-+ (NSError *)appAttestAttestKeyFailedWithError:(NSError *)error
-                                         keyId:(NSString *)keyId
-                                clientDataHash:(NSData *)clientDataHash;
-
-+ (NSError *)appAttestGenerateAssertionFailedWithError:(NSError *)error
-                                                 keyId:(NSString *)keyId
-                                        clientDataHash:(NSData *)clientDataHash;
-
-@end
-
-NS_ASSUME_NONNULL_END
+#import "_GACAppCheckErrorUtil.h"
